requirements.txt
django-model-utils>=2.0,<4.0
# Additional requirements go here
django-cities>=0.5
